Add files via upload
A neural net with an embedding layer. The first hidden layer of the network is an embedding layer. The second one is a fully connected layer. I have trained it on the imdb dataset from scratch. I used backprop eqns to compute the gradient values and gradient descent algorithm for optimizing the weight parameters.